+++ duckscript_sdk/src/sdk/std/test/assert_error/mod.rs
@@ -1,0 +1,40 @@
+use crate::utils::pckg;
+use duckscript::types::command::{Command, CommandResult};
+
+#[cfg(test)]
+#[path = "./mod_test.rs"]
+mod mod_test;
+
+struct CommandImpl {
+    package: String,
+}
+
+impl Command for CommandImpl {
+    fn name(&self) -> String {
+        pckg::concat(&self.package, "AssertError")
+    }
+
+    fn aliases(&self) -> Vec<String> {
+        vec!["assert_error".to_string()]
+    }
+
+    fn help(&self) -> String {
+        include_str!("help.md").to_string()
+    }
+
+    fn run(&self, arguments: Vec<String>) -> CommandResult {
+        let error_message = if arguments.is_empty() {
+            "Assert failed.".to_string()
+        } else {
+            arguments[0].clone()
+        };
+
+        CommandResult::Error(error_message)
+    }
+}
+
+pub(crate) fn create(package: &str) -> Box<dyn Command> {
+    Box::new(CommandImpl {
+        package: package.to_string(),
+    })
+}