+++ duckscript_sdk/src/sdk/std/test/assert_error/mod_test.rs
@@ -1,0 +1,17 @@
+use super::*;
+use crate::test;
+
+#[test]
+fn common_functions() {
+    test::test_common_command_functions(create(""));
+}
+
+#[test]
+fn run_no_args() {
+    test::run_script_and_error(vec![create("")], "out = assert_fail", "out");
+}
+
+#[test]
+fn run_with_message() {
+    test::run_script_and_error(vec![create("")], "out = assert_fail error", "out");
+}