+++ duckscript_sdk/src/sdk/std/test/mod.rs
@@ -1,5 +1,6 @@
 mod assert;
 mod assert_eq;
+mod assert_error;
 mod assert_fail;
 
 use crate::utils::pckg;
@@ -13,6 +14,7 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
 
     commands.set(assert::create(&package))?;
     commands.set(assert_eq::create(&package))?;
+    commands.set(assert_error::create(&package))?;
     commands.set(assert_fail::create(&package))?;
 
     Ok(())